requirement.txt
langchain-core
langchain-openai
openai
loguru
tqdm
python-dotenv
opencv-python
pillow
numpy